Performance: implement panama optimizations for L2LshModel
